Add support for defining schema types for reading data
This extends schema DSL with the ability to define "read" type that will
be used when a relation loads its tuples. By default we set this to a
noop passthrough-proc but if there are any attributes with a special
"read" type then we use relation's schema "relation hash" which is built
from schema attributes including "read" types.

This is a last-moment addition but it turned out we can't support
various custom DB types without this feature, and it was planned already
for rom 3.1.0 anyway.